requirements.txt
requests
reportlab
google-generativeai
python-dotenv
pillow
openai